hackone/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
pyserial
python-telegram-bot
